backend/rag/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
anthropic==0.33.0
python-dotenv==1.0.1
sentence-transformers==3.0.1
torch>=2.0.0
faiss-cpu>=1.9.0
numpy>=1.26.0
pypdf==4.2.0
markdown==3.6
tqdm==4.66.4
